resolve: load statements now create local bindings
Previously, load statements created global bindings.
As an often-unwanted consequence, names loaded by one module
would leak into its module dictionary and be visible to others.

This change to the resolver makes load define file-local variables
using a similar approach to Go: a new lexical block (the file block)
between function/comprehension blocks and the globals.
The set of names in the file block and the globals must be disjoint.

This incompatible change is guarded by the LoadBindsGlobally flag,
which is false by default. Users who want the old behavior may set
it to true, but they should prepare for it to go away entirely.
After that point, the old behavior may be simulated thus:
	- load("...", "x")
	+ load("...", _x="x"); x = _x # re-export

Observe that no run-time or API changes are required.

Change-Id: I5c91d40a0da1968ed34e5eb708cc5cd221141cd3